Run async Rx and Tx execution units in the same loop (#94)
Only use one asynchronous loop to run the Rx and Tx code, rather than one loop per execution unit. This should reduce contention in the AIM driver.

The biggest limitation is that the execution units still perform their own communication with the inline portion of the custom device. This opens up some possible data-tearing opportunities, where the inline portion of the custom device updates the FIFOs in-between execution units in the same loop.